Finished knight_travails function
This function receives two lists, the start square and the end square.
The objective is to count the number of moves a knight piece will do to
move from start to finish and return the path.

To do this, we have three classes, one for the board that will store in
an array the squares, that are objects from the Vertex class, that has
coords (x,y), a distance to measure the number of moves and parent, that
helps to get the path at the end of the function.

This function uses the BFS algorithm to determine the number of moves.
It creates the graph during the algorithm, creating new squares
(Vertex), calculating the next moves and checking if these moves are
valid. It finish when the current knight's position (a vertex) is the
desired square. Then it uses an auxiliary function to return the path
using a loop and adding each element to an array.